Rename _prox_grad to optax_optimistix_solvers
