+++ src/nemos/solvers/__init__.py
@@ -22,7 +22,7 @@ from .optimistix_solvers import (
 
 from ._solver_registry import solver_registry
 
-from ._prox_grad import (
+from .optax_optimistix_solvers import (
     OptaxOptimistixProximalGradient,
     OptaxOptimistixGradientDescent,
 )
